# openapi-3.0.yaml
openapi: 3.0.3
info:
  title: BNA REST API
  description: Provides a way to retrieve the BNA results programmatically.
  contact:
    name: The BNA Mechanics team
    url: https://peopleforbikes.github.io/
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development API
  - url: https://api.peopleforbikes.xyz
    description: Staging API
paths:
...
  /pipelines/bna:
    get:
      tags:
        - pipeline
      description: Get the details of all BNA pipelines
      operationId: get_pipelines_bnas
      parameters:
        - name: page_size
          in: query
          description: The number of items per page
          required: false
          schema:
            type: integer
            format: int64
            maximum: 65536
            minimum: 1
          example: 25
        - name: page
          in: query
          description: The result page being returned
          required: false
          schema:
            type: integer
            format: int64
            maximum: 65536
            minimum: 1
          example: 5
      responses:
        "200":
          description: Fetches the details of all BNA pipelines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BnaPipelines"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
    post:
      tags:
        - pipeline
      description: Create a new BNA pipeline
      operationId: post_pipelines_bna
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BnaPipelinePost"
        required: true
      responses:
        "201":
          description: Creates a new city submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BnaPipeline"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
  /pipelines/bna/{pipeline_id}:
    get:
      tags:
        - pipeline
      description: Get the details of a specific BNA pipeline
      operationId: get_pipelines_bna
      parameters:
        - name: pipeline_id
          in: path
          description: Pipeline identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fetches the details of a BNA pipeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BnaPipeline"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
    patch:
      tags:
        - pipeline
      description: Update the details of a specific BNA pipeline
      operationId: patch_pipelines_bna
      parameters:
        - name: pipeline_id
          in: path
          description: Pipeline identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BnaPipelinePatch"
        required: true
      responses:
        "200":
          description: Updates the details of a BNA pipeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BnaPipeline"
        "400":
          description: >-
            The request was formatted incorrectly or missing required
            parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      the request was formatted incorrectly or missing required
                      parameters
                    id: blfwkg8nvHcEJnQ=
                    source:
                      parameter: status
                    status: "400"
                    title: Bad Request
        "401":
          description: >-
            The request has not been fulfilled because it lacks valid
            authentication credentials for the target resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: >-
                      invalid authentication credentials to access the specified
                      resource
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "401"
                    title: Unauthorized
        "403":
          description: >-
            Forbidden to make the request. Most likely this indicates an issue
            with the credentials or permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: access to the requested resource is forbidden
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "403"
                    title: Forbidden
        "404":
          description: >-
            The particular resource requested was not found. This occurs, for
            example, when the id of the requested  resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrors"
              example:
                errors:
                  - details: the resource was not found
                    id: blfwkg8nvHcEJnQ=
                    source:
                      pointer: /bnas/ratings/e6aade5a-b343-120b-dbaa-bd916cd99221
                    status: "404"
                    title: Item Not Found
...
components:
  schemas:
    APIError:
      type: object
      description: >-
        Single API Error object as described in
        <https://jsonapi.org/format/#error-objects>.
      required:
        - status
        - title
        - details
        - source
      properties:
        details:
          type: string
          description: >-
            A human-readable explanation specific to this occurrence of the
            problem
          example: the entry was not found
        id:
          type: string
          description: A unique identifier for this particular occurrence of the problem.
          example: blfwkg8nvHcEJnQ=
          nullable: true
        source:
          $ref: "#/components/schemas/APIErrorSource"
          description: >-
            An object containing references to the primary source of the error.


            This field may be omitted in some situation. For instance, if the
            server cannot

            parse the request as valid JSON, including source doesn’t make sense

            (because there’s no JSON document for source to refer to).
        status:
          type: string
          description: >-
            The HTTP status code applicable to this problem, expressed as a
            string value.
          example: "404"
        title:
          type: string
          description: A short, human-readable summary of the problem
          example: Item Not Found
    APIErrorSource:
      oneOf:
        - type: object
          description: >-
            A JSON Pointer [RFC6901] to the value in the request document that
            caused the error.
          required:
            - pointer
          properties:
            pointer:
              type: string
              description: >-
                A JSON Pointer [RFC6901] to the value in the request document
                that caused the error.
          example:
            pointer: /data/attributes/title
        - type: object
          description: A string indicating which URI query parameter caused the error.
          required:
            - parameter
          properties:
            parameter:
              type: string
              description: A string indicating which URI query parameter caused the error.
          example:
            parameter: include
        - type: object
          description: >-
            A string indicating the name of a single request header which caused
            the error.
          required:
            - header
          properties:
            header:
              type: string
              description: >-
                A string indicating the name of a single request header which
                caused the error.
          example:
            header: Content-Type
      description: |-
        An object containing references to the primary source of the error.

        It SHOULD include one of the following members or be omitted:

          - pointer: a JSON Pointer [RFC6901](https://tools.ietf.org/html/rfc6901) to the
            value in the request document that caused the error [e.g. "/data" for a primary
            data object, or "/data/attributes/title" for a specific attribute].
            This MUST point to a value in the request document that exists; if it doesn’t,
            the client SHOULD simply ignore the pointer.
          - parameter: a string indicating which URI query parameter caused the error.
          - header: a string indicating the name of a single request header which caused the
            error.
    APIErrors:
      type: object
      description: >-
        Error objects MUST be returned as an array keyed by errors in the top
        level of a

        JSON:API document.
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/APIError"
    BnaPipeline:
      type: object
      required:
        - start_time
        - state_machine_id
        - status
        - step
      properties:
        cost:
          type: string
          description: Cost of an analysis in USD
          example: "6.8941"
          nullable: true
        end_time:
          type: string
          format: date-time
          description: End time
          nullable: true
        fargate_price_id:
          type: integer
          format: int32
          description: Fargate price identifier used to compute the cost
          nullable: true
        fargate_task_arn:
          type: string
          description: ARN of the Fargate task that performed the analysis
          example: >-
            arn:aws:ecs:us-west-2:123456789012:task/bna/29f979fc9fca402d94b014aa23d2f6e0
          nullable: true
        s3_bucket:
          type: string
          description: Path of the S3 bucket where the results were stored
          example: united states/new mexico/santa rosa/24.05.4
          nullable: true
        sqs_message:
          type: string
          description: Copy of the JSON message that was sent for processing
          example:
            city: santa rosa
            country: United States
            fips_code: "3570670"
            region: new mexico
          nullable: true
        start_time:
          type: string
          format: date-time
          description: Start time
        state_machine_id:
          type: string
          format: uuid
          description: >-
            Pipeline identifier

            This is the ID of the AWS state machine that was used to run the
            pipeline
        status:
          $ref: "#/components/schemas/PipelineStatus"
          description: Pipeline status
        step:
          $ref: "#/components/schemas/BnaPipelineStep"
          description: Last pipeline step that was completed
    BnaPipelinePatch:
      type: object
      required:
        - status
        - step
      properties:
        cost:
          type: string
          description: Cost of an analysis in USD
          example: "6.8941"
          nullable: true
        end_time:
          type: string
          format: date-time
          description: End time
          nullable: true
        fargate_price_id:
          type: integer
          format: int32
          description: Fargate price identifier used to compute the cost
          nullable: true
        fargate_task_arn:
          type: string
          description: ARN of the Fargate task that performed the analysis
          example: >-
            arn:aws:ecs:us-west-2:123456789012:task/bna/29f979fc9fca402d94b014aa23d2f6e0
          nullable: true
        s3_bucket:
          type: string
          description: Path of the S3 bucket where the results were stored
          example: united states/new mexico/santa rosa/24.05.4
          nullable: true
        sqs_message:
          type: string
          description: Copy of the JSON message that was sent for processing
          example:
            city: santa rosa
            country: United States
            fips_code: "3570670"
            region: new mexico
          nullable: true
        start_time:
          type: string
          format: date-time
          description: Start time
          nullable: true
        status:
          $ref: "#/components/schemas/PipelineStatus"
          description: Pipeline status
        step:
          $ref: "#/components/schemas/BnaPipelineStep"
          description: Last pipeline step that was completed
    BnaPipelinePost:
      type: object
      required:
        - state_machine_id
      properties:
        cost:
          type: string
          description: Cost of an analysis in USD
          example: "6.8941"
          nullable: true
        end_time:
          type: string
          format: date-time
          description: End time
          nullable: true
        fargate_price_id:
          type: integer
          format: int32
          description: Fargate price identifier used to compute the cost
          nullable: true
        fargate_task_arn:
          type: string
          description: ARN of the Fargate task that performed the analysis
          example: >-
            arn:aws:ecs:us-west-2:123456789012:task/bna/29f979fc9fca402d94b014aa23d2f6e0
          nullable: true
        s3_bucket:
          type: string
          description: Path of the S3 bucket where the results were stored
          example: united states/new mexico/santa rosa/24.05.4
          nullable: true
        sqs_message:
          type: string
          description: Copy of the JSON message that was sent for processing
          example:
            city: santa rosa
            country: United States
            fips_code: "3570670"
            region: new mexico
          nullable: true
        start_time:
          type: string
          format: date-time
          description: Start time
          nullable: true
        state_machine_id:
          type: string
          format: uuid
          description: >-
            Pipeline identifier

            This is the ID of the AWS state machine that was used to run the
            pipeline
    BnaPipelineStep:
      type: string
      enum:
        - SqsMessage
        - Setup
        - Analysis
        - Cleanup
    BnaPipelines:
      type: array
      items:
        $ref: "#/components/schemas/BnaPipeline"
...
    PipelineStatus:
      type: string
      enum:
        - Completed
        - Pending
        - Processing
...